requirements.txt
ifstools>=1.4
pycryptodome
kbinxml
tqdm